local2cloud/requirements.txt
boto3==1.26.100
google-cloud-storage==2.7.0
oauth2client==4.1.3
pytest==7.2.2
pytest-cov==4.0.0